Expanded unittests
Added more unit tests for the BFV scheme and serialisation. Linking them
all together so there is little repeated code. I purposefully returned
deep copies of each object to also ensure they were functional copies of
the original objects. Most of these tests will fail in the current state
of the repository, but once serialisation is complete (which is desired)
they will all test this functionality. There are still tests missing but
this is a good start.