Split unit test chain
The unit tests no longer chain from eachother but instead use core
functions to the test class. This means if an earlier one fails they
wont all fail, but they still validate the types and use of objects.
This will help better identify things that are unpickleable rather than
a dependancy of theirs that is not pickleable.