Added Latest Dockerfile and Updated Readme
Added Dockerfile build for huelse/seal-python of latest ubuntu version
with relevant documentation. I have also updated the residual
.dockerignore file to not ignore .git anymore since it will be necessary
to instantiate the submodules. I have modified the readme, taking out
the blank space and unecessary invisibles, while also adding notes on
dockerfile building and running. This can easily be extended to run
CI/CD tests in a dockerised environment in future if that is desirable.